+++ src/main/java/com/swen/news/global/service/ObjectStorageService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
+import software.amazon.awssdk.services.s3.model.ObjectCannedACL;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 
 import java.time.LocalDateTime;
@@ -52,6 +53,7 @@ public class ObjectStorageService {
                 .key(objectKey)
                 .contentType("audio/mpeg")
                 .contentLength((long) audioData.length)
+                    .acl(ObjectCannedACL.PUBLIC_READ)
                 .build();
             
             s3Client.putObject(putRequest, RequestBody.fromBytes(audioData));
